Working on Correlation Optimization
Adjusted a couple methods that dealt with correlation. Made them so that the didn't change the mutable portfolio dictionary. Basic correlation optimization that removes the stock from the portfolio with the largest weighted (by rank and Spearman rank coefficient) correlation. Still need to be able to plot the before/after distribution, backtest to see how the two different portfolios would perform. Need to change all correlation-related functions to correlate LOG RETURNS and NOT PRICE.